commit for modifications for db file

+++ db/migrate/20110623035041_create_gigavine_departments.rb
@@ -3,7 +3,7 @@ class CreateGigavineDepartments < ActiveRecord::Migration
     create_table :gigavine_departments do |t|
      t.string :name
      t.string :description
-     t.tinyint :status,:default => 1
+     t.smallint :status,:default => 1
      t.integer :author_id
 
       t.timestamps
